pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "openpixels"
version = "0.3.0"
authors = [
  { name="Robert", email="robert@openpixels.ai" },
]
description = "OpenPixels Python SDK"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = ["httpx>=0.28.1", "httpx[http2]>=0.28.1"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "python-dotenv"]

[project.urls]
Homepage = "https://openpixels.ai"
Issues = "https://github.com/openpixels-ai/openpixels-python/issues"